+++ src/app/services/LivestreamerService.js
@@ -85,9 +85,10 @@ define([
 	}
 
 
-	return Ember.Controller.extend( ChannelSettingsMixin, {
+	return Ember.Service.extend( ChannelSettingsMixin, {
 		metadata: Ember.inject.service(),
 		store   : Ember.inject.service(),
+		modal   : Ember.inject.service(),
 		settings: Ember.inject.service(),
 		chat    : Ember.inject.service(),
 
@@ -101,7 +102,7 @@ define([
 
 
 		startStream: function( stream, quality ) {
-			this.send( "openModal", "livestreamer", this, {
+			get( this, "modal" ).openModal( "livestreamer", this, {
 				error : null,
 				active: null,
 				abort : false
@@ -110,14 +111,22 @@ define([
 			var store   = get( this, "store" );
 			var channel = get( stream, "channel" );
 			var id      = get( channel, "id" );
+			var livestreamer;
 
 			// is the stream already running?
 			if ( store.hasRecordForId( "livestreamer", id ) ) {
-				return set( this, "active", store.recordForId( "livestreamer", id ) );
+				livestreamer = store.recordForId( "livestreamer", id );
+
+				if ( quality !== undefined && get( livestreamer, "quality" ) !== quality ) {
+					set( livestreamer, "quality", quality );
+				}
+
+				set( this, "active", livestreamer );
+				return;
 			}
 
 			// create a new livestreamer object
-			var livestreamer = store.createRecord( "livestreamer", {
+			livestreamer = store.createRecord( "livestreamer", {
 				id          : id,
 				stream      : stream,
 				channel     : channel,
@@ -156,12 +165,12 @@ define([
 
 			// automatically close modal on success
 			if ( get( this, "settings.gui_hidestreampopup" ) ) {
-				this.send( "close" );
+				get( this, "modal" ).closeModal();
 			}
 
 			// automatically open chat
 			if ( get( livestreamer, "gui_openchat" ) ) {
-				this.send( "chat", get( livestreamer, "channel" ) );
+				this.openChat( get( livestreamer, "channel" ) );
 			}
 
 			// hide the GUI
@@ -181,7 +190,7 @@ define([
 				   !get( livestreamer, "error" )
 				&& get( this, "active" ) === livestreamer
 			) {
-				this.send( "close" );
+				get( this, "modal" ).closeModal();
 			}
 
 			// restore the GUI
@@ -456,48 +465,10 @@ define([
 			}.bind( this ) );
 		},
 
-
-		actions: {
-			"download": function( success ) {
-				var url = get( this, "metadata.config.livestreamer-download-url" );
-				this.send( "openBrowser", url );
-				if ( success instanceof Function ) {
-					success();
-				}
-			},
-
-			"chat": function( channel ) {
-				var chat = get( this, "chat" );
-				chat.open( channel )
-					.catch(function(){});
-			},
-
-			"abort": function() {
-				set( this, "abort", true );
-				this.send( "closeModal" );
-			},
-
-			"close": function() {
-				this.send( "closeModal" );
-				run.schedule( "destroy", this, function() {
-					set( this, "active", null );
-				});
-			},
-
-			"shutdown": function() {
-				var active = get( this, "active" );
-				if ( active ) {
-					active.kill();
-				}
-				this.send( "close" );
-			},
-
-			"toggleLog": function() {
-				var active = get( this, "active" );
-				if ( active ) {
-					active.toggleProperty( "showLog" );
-				}
-			}
+		openChat: function( channel ) {
+			var chat = get( this, "chat" );
+			chat.open( channel )
+				.catch(function() {});
 		}
 	});